deepeval_demo/requirements.txt
deepeval>=0.12.0
langchain>=0.0.329
openai>=1.3.0
python-dotenv>=1.0.0
